Move version out of the helper package 🌿
